# PERF: Optimize SnakeCaseNamingPolicy.cs