@iigoshkin 2022-03-10
@perf-cat  Thank you for your work! I am very glad that we managed to improve the performance of this method